:recycle: Feat: Location 변환 로직 분리
<body>
- 기존 ScheduleConverter에 함께 있던 로직을 분리하고 스케줄 정보 컨버터 구현

- 관련 : #286
